+++ packages/core/test/mockWrapper.ts
@@ -59,6 +59,7 @@ const NATIVE: MockInterface<NativeType> = {
 
   crashedLastRun: jest.fn(),
   getNewScreenTimeToDisplay: jest.fn().mockResolvedValue(42),
+  getDataFromUri: jest.fn(),
 };
 
 NATIVE.isNativeAvailable.mockReturnValue(true);